requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
idna==3.4
jedi==0.18.2
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.8.0
tweepy==4.12.1
urllib3==1.26.13
wcwidth==0.2.5
yarg==0.1.9
